LayoutTests/picture/legitfails/media/picture-broken-video.html

<picture></picture>
<p>Test that an invalid picture file generates a MEDIA_ERR_SRC_NOT_SUPPORTED error and sets networkState to NETWORK_NO_SOURCE.</p>
<script src=resources/media-file.js></script>
<script src=resources/picture-test.js></script>

<script>
    picture.src = findMediaFile("picture", "content/garbage");
    waitForEvent("error", function () {
        testExpected("picture.error", "[object MediaError]");
        testExpected("picture.error.code", MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);
        testExpected("picture.networkState", HTMLMediaElement.NETWORK_NO_SOURCE);
        endTest();
    });
</script>
